requirements.txt
# Core requirements
numpy==2.0.0
gymnasium>=0.28.1
highway-env>=1.10.0
torch>=2.0.0
matplotlib>=3.7.0
pandas>=2.0.0
protobuf>=5.31.0
docopt>=0.6.2

# Video recording and display
pyvirtualdisplay>=3.0.0
opencv-python>=4.7.0
Pillow>=9.4.0

# Visualization
tqdm>=4.64.0
wandb>=0.15.0  # Optional for experiment tracking

# For development
pytest>=7.3.1  # Optional for testing
black>=23.3.0  # Optional for code formatting